+++ amostra/client/commands.py
@@ -6,10 +6,8 @@ from uuid import uuid4
 import requests
 import time as ttime
 from amostra.client import conf
-from os.path import expanduser
-import mongoquery
 
-# TODO: Separate local and online commands.py
+
 # TODO: Replace get, post, put and exception handling with _ from utils
 # TODO: Go over local lists prior to file lookups in local version
 # TODO: Make use of local lists for all collections
@@ -32,6 +30,7 @@ def _post(url, data):
 
 
 def _put(url, update_cont):
+    update_cont = {''}
     r = requests.put(url,
                      data=ujson.dumps(update_cont))
     r.raise_for_status()
@@ -46,9 +45,9 @@ class SampleReference(object):
         Parameters
         ----------
         host: str, optional
-            Machine name/address for amostra server
+            Machine name/address for Amostra server
         port: int, optional
-            Port amostra server is initiated on
+            Port Amostra server is initiated on
         """
         self.host = host
         self.port = port
@@ -85,10 +84,7 @@ class SampleReference(object):
                    name=name, time=time if time else ttime.time(),
                    container=container if container else 'NULL',
                    **kwargs)
-        domt = ujson.dumps(doc)
-        r = requests.post(self._samp_url,
-                          data=domt)
-        r.raise_for_status()
+        _get(self._samp_url,)
         return doc[uid]
 
     def create_sample_list(self, sample_list):
@@ -295,7 +291,7 @@ class ContainerReference(object):
         return payload
 
     def find(self, as_document=False, **kwargs):
-        """Given a set of mongo search parameters, return a requests iterator"""
+        """Given a set of MongoDB search parameters, return a requests iterator"""
         r = requests.get(self._cont_url,
                          params=ujson.dumps(kwargs))
         r.raise_for_status()
@@ -329,224 +325,3 @@ class ContainerReference(object):
                         'schema', params=ujson.dumps('container'))
         r.raise_for_status()
         return ujson.loads(r.text)
-
-
-def _find_local(fname, qparams):
-    """Find a document created using the local framework
-    Parameters
-    -----------
-    fname: str
-        Name of the query should be run
-    qparams: dict
-        Query parameters. Similar to online query methods
-
-    Yields
-    ------------
-    c: doct.Document, StopIteration
-        Result of the query if found
-
-    """
-    res_list = []
-    try:
-        with open(fname, 'r') as fp:
-            local_payload = ujson.load(fp)
-        qobj = mongoquery.Query(qparams)
-        for _sample in local_payload:
-            try:
-                qobj.match(_sample)
-                res_list.append(_sample)
-            except mongoquery.QueryError:
-                # this document is not it, skip it
-                pass
-    except FileNotFoundError:
-        raise RuntimeWarning('Local file {} does not exist'.format(fname))
-    for c in res_list:
-        yield c
-
-
-def _update_local(fname, qparams, replacement):
-    """Update a document created using the local framework
-    Parameters
-    -----------
-    fname: str
-        Name of the query should be run
-    qparams: dict
-        Query parameters. Similar to online query methods
-    replacement: dict
-        Fields/value pair to be updated. Beware of disallowed fields
-        such as time and uid
-    """
-    try:
-        with open(fname, 'r') as fp:
-            local_payload = ujson.load(fp)
-        qobj = mongoquery.Query(qparams)
-        for _sample in local_payload:
-            try:
-                qobj.match(_sample)
-                for k, v in replacement.iteritems():
-                    _sample[key] = v
-            except mongoquery.QueryError:
-                pass
-        with open(fname, 'w') as fp:
-            fp.write(ujson.dump(local_payload))
-    except FileNotFoundError:
-        raise RuntimeWarning('Local file {} does not exist'.format(fname))
-
-
-class LocalSampleReference:
-    """Handle sample information locally via json files"""
-    def __init__(self, top_dir=conf.local_conn_config['top']):
-        self.top_dir = top_dir
-        try:        
-            with open(self._samp_fname, 'r') as fp:
-                tmp = ujson.load(fp)
-        except FileNotFoundError:
-            tmp = []
-        try:
-            self.sample_list = tmp if tmp else []
-        except FileNotFoundError:
-            self.sample_list = []
-        
-    def create(self, name=None, time=None, uid=None, container=None,
-               **kwargs):
-        """Create a sample locally
-
-        Parameters
-        ----------
-        name: str
-            Name of the sample
-        time: float
-            Timestamp generated by the client
-        uid: str
-            Unique identifier for this sample
-        container: str, doct.Document
-            The container/group sample is contained within
-
-        Returns
-        -------
-        payload: dict
-            Document dict that was inserted
-        """
-        # TODO: Allow container to be an object
-        payload = dict(uid=uid if uid else str(uuid4()),
-                       name=name, time=time if time else ttime.time(),
-                       container=container if container else 'NULL',
-                       **kwargs)
-        self.sample_list.append(payload)
-        with open(self._samp_fname, 'w+') as fp:
-            ujson.dump(self.sample_list, fp)
-        return payload
-        
-    @property
-    def _samp_fname(self):
-        return expanduser(self.top_dir + '/samples.json')
-
-    def update(self, query, replacement):
-        _update_local(self._samp_fname, query, replacement)
-
-    def find(self, **kwargs):
-        return _find_local(self._samp_fname, kwargs)
-
-
-class LocalRequestReference:
-    """Handle request information locally via json files"""
-    def __init__(self, top_dir=conf.local_conn_config['top']):
-        self.top_dir = top_dir        
-        try:        
-            with open(self._req_fname, 'r') as fp:
-                tmp = ujson.load(fp)
-        except FileNotFoundError:
-            tmp = []
-        try:
-            self.request_list = tmp if tmp else []
-        except FileNotFoundError:
-            self.request_list = []
-    
-    @property
-    def _req_fname(self):
-        return expanduser(self.top_dir + '/requests.json')
-    
-    def create(self, sample=None, time=None, uid=None, state='active', 
-               seq_num=0, **kwargs):        
-        """Create a request locally
-
-        Parameters
-        ----------
-        sample: str, doct.Document, optional
-            Sample this request points to
-        time: float
-            Timestamp generated by the client
-        uid: str
-            Unique identifier for this sample
-        state: str
-            Defines whether this request is active or not
-        seq_num: int
-            Shows the order of request in the list of requests
-
-        Returns
-        -------
-        local_payload: dict
-            Request dict that was inserted locally
-
-        """
-        # TODO: Allow sample to be a doct
-        payload = dict(uid=uid if uid else str(uuid4()),
-                             sample=sample['uid'] if sample else 'NULL',
-                             time=time if time else ttime.time(), state=state,
-                             seq_num=seq_num, **kwargs)
-        self.request_list.append(payload)
-        with open(self._req_fname, 'w+') as fp:
-            ujson.dump(self.request_list, fp)
-        return payload
-    
-    def update(self, query, replacement):
-        _update_local(self._req_fname, query, replacement)
-
-
-    def find(self, **kwargs):
-        return _find_local(self._req_fname, kwargs)
-
-
-class LocalContainerReference:
-    """Handle container information locally via json files"""
-    def __init__(self, top_dir=conf.local_conn_config['top']):
-        self.top_dir = top_dir
-        try:        
-            with open(self._cont_fname, 'r') as fp:
-                tmp = ujson.load(fp)
-        except FileNotFoundError:
-            tmp = []
-        self.container_list = tmp if tmp else []
-        self.container_list = []
-        
-    @property
-    def _cont_fname(self):
-        return expanduser(self.top_dir + '/containers.json')
-
-    def create(self, uid=None, time=None, **kwargs):
-        """ Create a container locally.
-
-        Parameters
-        ----------
-        time: float
-            Timestamp generated by the client
-        uid: str
-            Unique identifier for this sample
-
-        Returns
-        -------
-        payload: dict
-            Document dict that was inserted
-        """
-        payload = dict(uid=uid if uid else str(uuid4()),
-                       time=time if time else ttime.time(), **kwargs)
-        self.container_list.append(payload)
-        with open(self._cont_fname, 'w+') as fp:
-            ujson.dump(self.container_list, fp)
-        return payload
-
-    def update(self, query, replacement):
-        _update_local(self._cont_fname, query, replacement)
-
-    def find(self, **kwargs):
-        return _find_local(self._cont_fname, kwargs)
